+++ packages/std/src/results/mod.rs
@@ -4,10 +4,9 @@ mod attribute;
 mod context;
 mod contract_result;
 mod cosmos_msg;
-mod handle;
-mod init;
-mod migrate;
+mod empty;
 mod query;
+mod response;
 mod system_result;
 
 pub use attribute::{attr, Attribute};
@@ -15,8 +14,11 @@ pub use attribute::{attr, Attribute};
 pub use context::Context;
 pub use contract_result::ContractResult;
 pub use cosmos_msg::{wasm_execute, wasm_instantiate, BankMsg, CosmosMsg, StakingMsg, WasmMsg};
-pub use handle::HandleResponse;
-pub use init::InitResponse;
-pub use migrate::MigrateResponse;
+pub use empty::Empty;
 pub use query::QueryResponse;
+pub use response::Response;
 pub use system_result::SystemResult;
+
+pub type InitResponse<T = Empty> = Response<T>;
+pub type HandleResponse<T = Empty> = Response<T>;
+pub type MigrateResponse<T = Empty> = Response<T>;